requirements.txt
numpy==1.13.3
matplotlib==2.1.2
pandas==0.22.0
Pillow==5.1.0
beautifulsoup4==4.6.0
